Modify use cases to run on WGR development board description:Modify use cases to run on WGR development board issue:https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I5QD96 Signed-off-by: jing_han <hanjing35@huawei.com>

+++ test/fuzztest/stopheaptracking1_fuzzer/stopheaptracking1_fuzzer.cpp
@@ -29,6 +29,9 @@ namespace OHOS {
         RuntimeOption option;
         option.SetLogLevel(RuntimeOption::LOG_LEVEL::ERROR);
         EcmaVM *vm = JSNApi::CreateJSVM(option);
+        if (size <= 0) {
+            return;
+        }
         std::string path(data, data + size);
         DFXJSNApi::StopHeapTracking(vm, path);
         JSNApi::DestroyJSVM(vm);
